algorithm.montecarlo.AlgorithmBase
